[#44] feat: observing auth and effecting on myInfo
